# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Open Card Control Plane API
  version: 1.1
  description: >-
    REST JSON and persistent SSE control-plane endpoints. Product state remains owned by the
    control plane; provider and agent endpoints are versioned contracts. API clients may send
    an exact Open-Card-API-Version header or an Open-Card-API-Min-Version and
    Open-Card-API-Max-Version range. All version headers are optional for backwards
    compatibility: clients that omit them negotiate the legacy 1.0 contract. Requests with
    no common supported version receive 426 Upgrade Required.
servers:
  - url: /
paths:
...
  /api/v1/applications:
    parameters:
      - $ref: '#/components/parameters/APIVersion'
      - $ref: '#/components/parameters/APIMinVersion'
      - $ref: '#/components/parameters/APIMaxVersion'
    get:
      operationId: listApplications
      responses:
        '200':
          description: Applications
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
    post:
      operationId: createApplication
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Created
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApplicationResponse'
        '400':
          description: Invalid request
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '409':
          description: Idempotency key was reused with different input
          headers:
            Open-Card-API-Version:
              $ref: '#/components/headers/APINegotiatedVersion'
        '426':
          $ref: '#/components/responses/APIUpgradeRequired'
...
  /api/v1/operations/views/{applicationId}:
    get:
      operationId: getApplicationOperationsView
      parameters:
        - $ref: '#/components/parameters/ApplicationID'
        - {name: environment_id, in: query, required: true, schema: {type: string}}
        - {name: mode, in: query, required: false, schema: {type: string, enum: [ordinary, operator], default: ordinary}}
        - {name: X-Open-Card-Role, in: header, required: false, schema: {type: string, enum: [operator]}}
      responses:
        '200':
          description: Both projections carry one control-plane fact version and observation time.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationsViewResponse'}
        '403': {description: Operator projection requires operator role}
        '503': {description: Stored operation facts are unavailable or invalid}
...
components:
  parameters:
    APIVersion:
      name: Open-Card-API-Version
      in: header
      required: false
      description: >-
        Request an exact API version. This header is mutually exclusive with the minimum and
        maximum version range headers. If all version headers are omitted, API version 1.0 is
        negotiated for legacy clients.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    APIMinVersion:
      name: Open-Card-API-Min-Version
      in: header
      required: false
      description: Lowest API version accepted by the client. Use with Open-Card-API-Max-Version.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    APIMaxVersion:
      name: Open-Card-API-Max-Version
      in: header
      required: false
      description: Highest API version accepted by the client. Use with Open-Card-API-Min-Version.
      schema:
        type: string
        pattern: '^v?[0-9]+(\.[0-9]+)?$'
    ApplicationID:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Stable retry key. Reuse with different input is rejected.
      schema:
        type: string
        minLength: 1
...
  headers:
    APINegotiatedVersion:
      description: API version selected for this request and response.
      schema:
        type: string
        pattern: '^[0-9]+\.[0-9]+$'
  responses:
    APIUpgradeRequired:
      description: No common API version is available for the requested range.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    APIError:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
    Application:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApplicationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Application'
    CreateApplicationResponse:
      type: object
      required: [application, operation_id]
      properties:
        application:
          $ref: '#/components/schemas/Application'
        operation_id:
          type: string
    CreateApplicationRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
...
    ResourceObservation:
      type: object
      additionalProperties: false
      required: [cpu_cores, memory_bytes, disk_bytes, network_rx_bytes, network_tx_bytes, restart_count, pids_current, applied_cpu_millicores, applied_memory_bytes, applied_pids, changed_path_count, cgroup_verified, metrics_known]
      properties:
        container_id: {type: string}
        cpu_cores: {type: number, minimum: 0}
        memory_bytes: {type: integer, format: int64, minimum: 0}
        disk_bytes: {type: integer, format: int64, minimum: 0}
        network_rx_bytes: {type: integer, format: int64, minimum: 0}
        network_tx_bytes: {type: integer, format: int64, minimum: 0}
        restart_count: {type: integer, format: int64, minimum: 0}
        pids_current: {type: integer, format: int64, minimum: 0}
        applied_cpu_millicores: {type: integer, format: int64, minimum: 0}
        applied_memory_bytes: {type: integer, format: int64, minimum: 0}
        applied_pids: {type: integer, format: int64, minimum: 0}
        changed_path_count: {type: integer, format: int64, minimum: 0}
        cgroup_verified: {type: boolean}
        metrics_known: {type: boolean}
        host_port: {type: integer, minimum: 0, maximum: 65535}
    ServiceOperationsFact:
      type: object
      additionalProperties: false
      required: [name, role, deployment_id, release_id, status, healthy, required, impact, actual, observed_at]
      properties:
        name: {type: string}
        role: {type: string, enum: [ingress, worker, stateful, one_shot]}
        deployment_id: {type: string}
        release_id: {type: string}
        status: {type: string}
        healthy: {type: boolean}
        required: {type: boolean}
        impact: {type: string}
        next_action: {type: string}
        actual: {$ref: '#/components/schemas/ResourceObservation'}
        observed_at: {type: string, format: date-time}
    ApplicationOperationsView:
      type: object
      additionalProperties: false
      required: [version, application_id, environment_id, release_id, state, summary, services, allowed_actions, data_notice, ai_status, observed_at]
      properties:
        version: {type: string}
        application_id: {type: string}
        environment_id: {type: string}
        release_id: {type: string}
        state: {type: string, enum: [healthy, attention, partial, stopped]}
        summary: {type: string}
        services: {type: array, items: {$ref: '#/components/schemas/ServiceOperationsFact'}}
        allowed_actions:
          type: object
          required: [restart_services, redeploy, rollback, rollback_data]
          properties:
            restart_services: {type: array, items: {type: string}}
            redeploy: {type: boolean}
            rollback: {type: boolean}
            rollback_data: {type: boolean, enum: [false]}
        data_notice: {type: string}
        ai_status: {type: string, enum: [disabled, unavailable]}
        observed_at: {type: string, format: date-time}
    OrdinaryOperationsView:
      type: object
      additionalProperties: false
      required: [version, application_id, state, summary, impacts, next_actions, observed_at]
      properties:
        version: {type: string}
        application_id: {type: string}
        state: {type: string, enum: [healthy, attention, partial, stopped]}
        summary: {type: string}
        impacts: {type: array, items: {type: string}}
        next_actions: {type: array, items: {type: string}}
        observed_at: {type: string, format: date-time}
    OperationsViewResponse:
      type: object
      required: [mode, facts]
      properties:
        mode: {type: string, enum: [ordinary, operator]}
        facts:
          oneOf:
            - {$ref: '#/components/schemas/OrdinaryOperationsView'}
            - {$ref: '#/components/schemas/ApplicationOperationsView'}
...